Move builders from ActivityInput to MessageActivityInput  (#363)
- only outbound activities we have are `MessageActivityInput` and
`TypingActivityInput`
- basically same as #231, moved citations, feedback, AI label

---------
