+++ packages/api/src/microsoft_teams/api/activities/message/message.py
@@ -22,6 +22,17 @@ from ...models import (
     SuggestedActions,
     TextFormat,
 )
+from ...models.channel_data import FeedbackLoop
+from ...models.entity import (
+    AIMessageEntity,
+    Appearance,
+    CitationAppearance,
+    CitationEntity,
+    Claim,
+    Entity,
+    Image,
+    MessageEntity,
+)
 from ..utils import StripMentionsTextOptions, strip_mentions_text
 
 
@@ -323,6 +334,85 @@ class MessageActivityInput(_MessageBase, ActivityInputBase):
         stream_entity = StreamInfoEntity(type="streaminfo", stream_id=self.id, stream_type="final")
 
         return self.add_entity(stream_entity)
+
+    def add_ai_generated(self) -> Self:
+        """Add the 'Generated By AI' label."""
+        message_entity = self._ensure_single_root_level_message_entity()
+        ai_entity = AIMessageEntity(**message_entity.model_dump())
+        if ai_entity.additional_type and "AIGeneratedContent" in ai_entity.additional_type:
+            return self
+
+        if not ai_entity.additional_type:
+            ai_entity.additional_type = []
+
+        ai_entity.additional_type.append("AIGeneratedContent")
+
+        self._update_entity(message_entity, ai_entity)
+
+        return self
+
+    def add_citation(self, position: int, appearance: CitationAppearance) -> Self:
+        """Add citations."""
+        message_entity = self._ensure_single_root_level_message_entity()
+        citation_entity = CitationEntity(**message_entity.model_dump())
+        if citation_entity.citation is None:
+            citation_entity.citation = []
+
+        citation_entity.citation.append(
+            Claim(
+                position=position,
+                appearance=Appearance(
+                    abstract=appearance.abstract,
+                    name=appearance.name,
+                    image=Image(name=appearance.icon) if appearance.icon else None,
+                    keywords=appearance.keywords,
+                    text=appearance.text,
+                    url=appearance.url,
+                    usage_info=appearance.usage_info,
+                ),
+            )
+        )
+
+        self._update_entity(message_entity, citation_entity)
+
+        return self
+
+    def _ensure_single_root_level_message_entity(self) -> MessageEntity:
+        """
+        Get or create the base message entity.
+        There should only be one root level message entity.
+        """
+        message_entity = next(
+            (e for e in (self.entities or []) if e.type == "https://schema.org/Message" and e.at_type == "Message"),
+            None,
+        )
+
+        if not message_entity:
+            message_entity = MessageEntity()
+            self.add_entity(message_entity)
+
+        return message_entity
+
+    def _update_entity(self, old_entity: Entity, new_entity: Entity) -> None:
+        if self.entities is not None:
+            index = self.entities.index(old_entity)
+            self.entities.pop(index)
+            self.entities.insert(index, new_entity)
+
+    def add_feedback(self, mode: Literal["default", "custom"] = "default") -> Self:
+        """
+        Enable message feedback.
+
+        Args:
+            mode: "default" shows Teams' built-in thumbs up/down UI.
+                  "custom" triggers a message/fetchTask invoke so the bot
+                  can return its own task module dialog.
+        """
+        if not self.channel_data:
+            self.channel_data = ChannelData()
+        self.channel_data.feedback_loop = FeedbackLoop(type=mode)
+        self.channel_data.feedback_loop_enabled = None
+        return self
 
     def with_recipient(self, value: Account, is_targeted: Optional[bool] = None) -> Self:
         """
