+++ packages/api/src/microsoft_teams/api/models/activity.py
@@ -5,24 +5,15 @@ Licensed under the MIT License.
 
 import warnings
 from datetime import datetime
-from typing import Any, List, Literal, Optional, Self
+from typing import Any, List, Optional, Self
 
 from microsoft_teams.api.models.account import Account, ConversationAccount
-from microsoft_teams.api.models.channel_data.channel_data import ChannelData, FeedbackLoop
+from microsoft_teams.api.models.channel_data.channel_data import ChannelData
 from microsoft_teams.api.models.channel_data.channel_info import ChannelInfo
 from microsoft_teams.api.models.channel_data.notification_info import NotificationInfo
 from microsoft_teams.api.models.channel_data.team_info import TeamInfo
 from microsoft_teams.api.models.channel_id import ChannelID
-from microsoft_teams.api.models.entity.ai_message_entity import AIMessageEntity
-from microsoft_teams.api.models.entity.citation_entity import (
-    Appearance,
-    CitationAppearance,
-    CitationEntity,
-    Claim,
-    Image,
-)
 from microsoft_teams.api.models.entity.entity import Entity
-from microsoft_teams.api.models.entity.message_entity import MessageEntity
 from microsoft_teams.api.models.meetings.meeting_info import MeetingInfo
 from microsoft_teams.common.experimental import ExperimentalWarning
 
@@ -213,88 +204,9 @@ class ActivityInput(_ActivityBase):
         self.entities.extend(values)
         return self
 
-    def add_ai_generated(self) -> Self:
-        """Add the 'Generated By AI' label."""
-        message_entity = self.ensure_single_root_level_message_entity()
-        ai_entity = AIMessageEntity(**message_entity.model_dump())
-        if ai_entity.additional_type and "AIGeneratedContent" in ai_entity.additional_type:
-            return self
-
-        if not ai_entity.additional_type:
-            ai_entity.additional_type = []
-
-        ai_entity.additional_type.append("AIGeneratedContent")
-
-        self._update_entity(message_entity, ai_entity)
-
-        return self
-
-    def add_feedback(self, mode: Literal["default", "custom"] = "default") -> Self:
-        """
-        Enable message feedback.
-
-        Args:
-            mode: "default" shows Teams' built-in thumbs up/down UI.
-                  "custom" triggers a message/fetchTask invoke so the bot
-                  can return its own task module dialog.
-        """
-        if not self.channel_data:
-            self.channel_data = ChannelData()
-        self.channel_data.feedback_loop = FeedbackLoop(type=mode)
-        self.channel_data.feedback_loop_enabled = None
-        return self
-
-    def add_citation(self, position: int, appearance: CitationAppearance) -> Self:
-        """Add citations."""
-        message_entity = self.ensure_single_root_level_message_entity()
-        citation_entity = CitationEntity(**message_entity.model_dump())
-        if citation_entity.citation is None:
-            citation_entity.citation = []
-
-        citation_entity.citation.append(
-            Claim(
-                position=position,
-                appearance=Appearance(
-                    abstract=appearance.abstract,
-                    name=appearance.name,
-                    image=Image(name=appearance.icon) if appearance.icon else None,
-                    keywords=appearance.keywords,
-                    text=appearance.text,
-                    url=appearance.url,
-                    usage_info=appearance.usage_info,
-                ),
-            )
-        )
-
-        self._update_entity(message_entity, citation_entity)
-
-        return self
-
     def is_streaming(self) -> bool:
         """Check if this is a streaming activity."""
         return bool(self.entities and any(e.type == "streaminfo" for e in self.entities or []))
-
-    def ensure_single_root_level_message_entity(self) -> MessageEntity:
-        """
-        Get or create the base message entity.
-        There should only be one root level message entity.
-        """
-        message_entity = next(
-            (e for e in (self.entities or []) if e.type == "https://schema.org/Message" and e.at_type == "Message"),
-            None,
-        )
-
-        if not message_entity:
-            message_entity = MessageEntity()
-            self.add_entity(message_entity)
-
-        return message_entity
-
-    def _update_entity(self, old_entity: Entity, new_entity: Entity) -> None:
-        if self.entities is not None:
-            index = self.entities.index(old_entity)
-            self.entities.pop(index)
-            self.entities.insert(index, new_entity)
 
 
 class Activity(_ActivityBase):
